Add @phosphor-icons/react package and update Button component
The @phosphor-icons/react package has been added to the dependencies in package.json and package-lock.json files. This new dependency is used in the Button component, where a PlusCircle icon from this package is utilized. The CSS for the Button component has also been adjusted to accommodate this new icon.