Implement task status toggle functionality
Added the ability to alter the status of a task from unfinished to finished and vice versa in Item.tsx component. Incorporated useState to manage task status state, functions to render different styles based on whether a task is done or not, and an onClick event to change event status. Modified corresponding styles in Item.module.css to support this functionality.